backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
pymongo==4.6.1
redis==5.0.1
bcrypt==4.1.2
